Move more exception handling into the BaseFuture
